+++ R/utils.R
@@ -162,23 +162,18 @@ r_to_py.error <- function(x, convert = FALSE) {
   e
 }
 
-#' @export
-conditionCall.python.builtin.BaseException <- function(c) {
-  as_r_value(py_get_attr(c, "call", TRUE))
-}
-
-#' @export
-conditionMessage.python.builtin.BaseException <- function(c) {
-  conditionMessage_from_py_exception(c)
-}
-
 
 #' @export
 `$.python.builtin.BaseException` <- function(x, name) {
-  if(identical(name, "call"))
-    return(conditionCall(x))
-  if(identical(name, "message"))
-    return(conditionMessage(x))
+  if(identical(name, "call")) {
+    out <- if(typeof(x) == "list") unclass(x)[["call"]]
+    return(out %||% as_r_value(py_get_attr(x, "call", TRUE)))
+  }
+
+  if(identical(name, "message")) {
+    out <- if(typeof(x) == "list") unclass(x)[["message"]]
+    return(out %||% conditionMessage_from_py_exception(x))
+  }
 
   py_maybe_convert(py_get_attr(x, name, TRUE), py_has_convert(x))
 }