+++ src/python.cpp
@@ -606,7 +606,7 @@ std::string as_r_class(PyObject* classPtr) {
 
 }
 
-SEXP py_class_names(PyObject* object) {
+SEXP py_class_names(PyObject* object, bool exception) {
 
   // Py_TYPE() usually returns a borrowed reference to object.__class__
   // but can differ if __class__ was modified after the object was created.
@@ -666,7 +666,7 @@ SEXP py_class_names(PyObject* object) {
     classNames.insert(classNames.end() - 1, "python.builtin.iterator");
 
   // if it's a BaseException instance, append "error"/"interrupt" and "condition"
-  if (PyExceptionInstance_Check(object)) {
+  if (exception) {
     if (PyErr_GivenExceptionMatches(type, PyExc_KeyboardInterrupt))
       classNames.push_back("interrupt");
     else
@@ -675,7 +675,13 @@ SEXP py_class_names(PyObject* object) {
   }
 
   RObject classNames_robj = Rcpp::wrap(classNames); // convert + protect
-  return eval_call(r_func_py_filter_classes, (SEXP) classNames_robj);
+  RObject out = eval_call(r_func_py_filter_classes, (SEXP) classNames_robj);
+  return out;
+}
+
+
+SEXP py_class_names(PyObject* object) {
+  return py_class_names(object, (bool) PyExceptionInstance_Check(object));
 }
 
 
@@ -719,6 +725,27 @@ bool inherits2(SEXP object, const char* name) {
     for (int i = Rf_length(klass)-1; i >= 0; i--) {
       if (strcmp(CHAR(STRING_ELT(klass, i)), name) == 0)
         return true;
+    }
+  }
+  return false;
+}
+
+bool inherits2(SEXP object, const char* name1, const char* name2) {
+  // like inherits in R, but iterates over the class STRSXP vector
+  // in reverse, since python.builtin.object is typically at the tail.
+  SEXP klass = Rf_getAttrib(object, R_ClassSymbol);
+  if (TYPEOF(klass) == STRSXP) {
+
+    int i = Rf_length(klass)-1;
+
+    for (; i >= 0; i--) {
+      if (strcmp(CHAR(STRING_ELT(klass, i)), name2) == 0) {
+        // found name2, now look for name1
+        for (i--; i >= 0; i--)
+          if (strcmp(CHAR(STRING_ELT(klass, i)), name1) == 0)
+            return true; // found name1 also
+        break; // did not find name1
+      }
     }
   }
   return false;
@@ -891,10 +918,10 @@ SEXP py_fetch_error(bool maybe_reuse_cached_r_trace) {
   // for shallow call stacks. So we fetch the call directly
   // using the R API.
   if (!PyObject_HasAttrString(excValue, "call")) {
-    // Technically we don't need to protct call, since
+    // Technically we don't need to protect call, since
     // it would already be protected by it's inclusion in the R callstack,
     // but rchk flags it anyway, and so ...
-    RObject r_call(  get_current_call() );
+    RObject r_call( get_current_call() );
     PyObject* r_call_capsule(py_capsule_new(r_call));
     PyObject_SetAttrString(excValue, "call", r_call_capsule);
     Py_DecRef(r_call_capsule);
@@ -943,22 +970,35 @@ class PyFlushOutputOnScopeExit {
 };
 
 
-// [[Rcpp::export]]
-std::string conditionMessage_from_py_exception(PyObjectRef exc) {
-  // invoke 'traceback.format_exception_only(<traceback>)'
-  PyObjectPtr tb_module(py_import("traceback"));
-  if (tb_module.is_null())
-    return "<unknown python exception, traceback module not found>";
 
-  PyObjectPtr format_exception_only(
-      PyObject_GetAttrString(tb_module, "format_exception_only"));
-  if (format_exception_only.is_null())
-    return "<unknown python exception, traceback format fn not found>";
+std::string conditionMessage_from_py_exception(PyObject* exc) {
+  // invoke 'traceback.format_exception_only(<traceback>)'
+
+  static PyObject* format_exception_only = []() {
+    PyObjectPtr tb_module(py_import("traceback"));
+    if (tb_module.is_null()) {
+      PyErr_Print();
+      Rcpp::stop("Failed to format Python Exception; could not import traceback module");
+    }
+
+    PyObject* format_exception_only = PyObject_GetAttrString(tb_module, "format_exception_only");
+
+    if (format_exception_only == NULL) {
+      PyErr_Print();
+      Rcpp::stop("Failed to format Python Exception; could not get traceback.format_exception_only");
+    }
+
+    return format_exception_only;
+  }();
+
 
   PyObjectPtr formatted(PyObject_CallFunctionObjArgs(
-      format_exception_only, Py_TYPE(exc.get()), exc.get(), NULL));
-  if (formatted.is_null())
-    return "<unknown python exception, traceback format fn returned NULL>";
+      format_exception_only, Py_TYPE(exc), exc, NULL));
+
+  if (formatted.is_null()) {
+    PyErr_Print();
+    Rcpp::stop("Failed to format Python Exception; traceback.format_exception_only() raised an Exception");
+  }
 
   // build error text
   std::ostringstream oss;
@@ -967,14 +1007,12 @@ std::string conditionMessage_from_py_exception(PyObjectRef exc) {
   for (Py_ssize_t i = 0, n = PyList_Size(formatted); i < n; i++)
     oss << as_std_string(PyList_GetItem(formatted, i));
 
-  static std::string hint;
-
-  if (hint.empty()) {
+  static std::string hint = []() {
     Environment pkg_env(Environment::namespace_env("reticulate"));
     Function hint_fn = pkg_env[".py_last_error_hint"];
     CharacterVector r_result = hint_fn();
-    hint = Rcpp::as<std::string>(r_result[0]);
-  }
+    return Rcpp::as<std::string>(r_result[0]);
+  }();
 
   oss << hint;
   std::string error = oss.str();
@@ -991,7 +1029,7 @@ std::string conditionMessage_from_py_exception(PyObjectRef exc) {
 
     std::string trunc("<...truncated...>");
 
-    // Tensorflow since ~2.6 has been including a currated traceback as part of
+    // TensorFlow since ~2.6 has been including a curated traceback as part of
     // the formatted exception message, with the most user-actionable content
     // towards the tail. Since the tail is the most useful part of the message,
     // truncate from the middle of the exception by default, after including the
@@ -1008,6 +1046,11 @@ std::string conditionMessage_from_py_exception(PyObjectRef exc) {
   }
 
   return error;
+}
+
+// [[Rcpp::export]]
+std::string conditionMessage_from_py_exception(PyObjectRef exc) {
+  return conditionMessage_from_py_exception(exc.get());
 }
 
 // check whether the PyObject can be mapped to an R scalar type
@@ -1223,6 +1266,8 @@ bool is_py_object(SEXP x) {
     case ENVSXP:
     case CLOSXP:
       return inherits2(x, "python.builtin.object");
+    case VECSXP:
+      return inherits2(x, "python.builtin.object", "condition");
     }
   }
   return false;
@@ -4374,4 +4419,28 @@ bool try_py_resolve_module_proxy(SEXP proxy) {
   Rcpp::Environment pkgEnv = Rcpp::Environment::namespace_env("reticulate");
   Rcpp::Function py_resolve_module_proxy = pkgEnv["py_resolve_module_proxy"];
   return py_resolve_module_proxy(proxy);
+}
+
+
+
+SEXP py_exception_as_condition(PyObject* object, SEXP refenv) {
+  static SEXP names = []() {
+    SEXP names = Rf_allocVector(STRSXP, 2);
+    R_PreserveObject(names);
+    SET_STRING_ELT(names, 0, Rf_mkChar("message"));
+    SET_STRING_ELT(names, 1, Rf_mkChar("call"));
+    return names;
+  }();
+  SEXP out = PROTECT(Rf_allocVector(VECSXP, 2));
+
+  SET_VECTOR_ELT(out, 0, Rcpp::wrap(conditionMessage_from_py_exception(object)));
+  PyObject* call = py_get_attr(object, "call");
+  if(call != NULL)
+    SET_VECTOR_ELT(out, 1, py_to_r(call, true));
+
+  Rf_setAttrib(out, R_NamesSymbol, names);
+  Rf_setAttrib(out, R_ClassSymbol, Rf_getAttrib(refenv, R_ClassSymbol));
+  Rf_setAttrib(out, sym_py_object, refenv);
+  UNPROTECT(1);
+  return out;
 }